requirements.txt
# basis
requests==2.31.0
envparse==0.2.0

# json validation (inspector)
jsonschema==4.21.1

# for version comparison
semver==3.0.2

# testing
pytest==8.1.0
pytest-xdist[psutil]==3.5.0
betamax==0.8.1
coverage~=7.4.3  # pinned to <7 as coveralls is preventing v7 for coverage
pytest-cov==4.1.0
coveralls==3.3.1
Pillow==10.0.1

# documentation
Sphinx==7.2.6; python_version >= '3.9'
Sphinx~=7.1.2; python_version <= '3.8'
nbsphinx==0.9.3

# integration
tox==4.12.1

# code style
flake8==7.0.0
importlib-metadata==7.0.1
pydocstyle==6.3.0
mypy==1.8.0; python_version >= '3.4'
pre-commit==3.6.2

# for package management (uploads)
twine~=4.0; python_version >= '3.7'
twine~=3.0; python_version <= '3.6'

pytz==2024.1
